requirements.txt
tensorflow
numpy
matplotlib
Pillow
pyyaml